Move check for non-noneness of gurobi license file into install step

+++ easybuild/easyblocks/g/gurobi.py
@@ -60,9 +60,6 @@ class EB_Gurobi(Tarball):
         if self.orig_license_file is None:
             self.orig_license_file = os.getenv('EB_GUROBI_LICENSE_FILE', None)
 
-        if self.orig_license_file is None or not os.path.exists(self.cfg['license_file']):
-            raise EasyBuildError("No existing license file specified: %s", self.orig_license_file)
-
         if self.cfg['copy_license_file']:
             self.license_file = os.path.join(self.installdir, 'gurobi.lic')
         else:
@@ -73,6 +70,9 @@ class EB_Gurobi(Tarball):
         super(EB_Gurobi, self).install_step()
 
         if self.cfg['copy_license_file']:
+            if self.orig_license_file is None or not os.path.exists(self.cfg['license_file']):
+                raise EasyBuildError("No existing license file specified: %s", self.orig_license_file)
+
             copy_file(self.orig_license_file, self.license_file)
 
         if get_software_root('Python'):
